+++ packages/web3-providers/src/Fuse/apis/Base.ts
@@ -1,5 +1,5 @@
 import Fuse from 'fuse.js'
-import type { FuseBaseAPI } from '../../types/Fuse.js'
+import type { FuseBaseAPI } from '../../entry-types.js'
 
 export class FuseAPI implements FuseBaseAPI.Provider {
     create<T>(items: T[], options?: Fuse.IFuseOptions<T>, index?: Fuse.FuseIndex<T>) {
