+++ packages/web3-providers/src/SmartPay/apis/FunderAPI.ts
@@ -3,9 +3,9 @@ import { BigNumber } from 'bignumber.js'
 import { ChainId, type TransactionReceipt } from '@masknet/web3-shared-evm'
 import { EMPTY_LIST, type Proof } from '@masknet/shared-base'
 import { Web3API } from '../../Connection/index.js'
-import { FunderAPI } from '../../types/Funder.js'
-import { fetchJSON } from '../../entry-helpers.js'
 import { FUNDER_PROD } from '../constants.js'
+import { FunderAPI } from '../../entry-types.js'
+import { fetchJSON } from '../../entry-helpers.js'
 
 export class SmartPayFunderAPI implements FunderAPI.Provider<ChainId> {
     private web3 = new Web3API()
